pyproject.toml
[project]
name = "agentgemini"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "openai-agents>=0.6.1",
    "pypdf2>=3.0.1",
    "python-dotenv>=1.2.1",
    "requests",
    "google-generativeai",
    "streamlit",
    "agent>=0.1.3",
]

[tool.setuptools]
py-modules = ["agent", "app", "main", "tools"]

